# backend/swagger.json
{
	"basePath": "/api",
	"consumes": [
		"application/json"
	],
	"definitions": {
		"IUser": {
			"properties": {
				"email": {
					"type": "string"
				},
				"role": {
					"type": "number",
					"format": "double"
				},
				"name": {
					"type": "string"
				},
				"surname": {
					"type": "string"
				},
				"patronymic": {
					"type": "string",
					"x-nullable": true
				},
				"password": {
					"type": "string"
				}
			},
			"required": [
				"email",
				"role",
				"name",
				"surname",
				"password"
			],
			"type": "object",
			"additionalProperties": false
		},
		"AuthCred": {
			"properties": {
				"email": {
					"type": "string"
				},
				"password": {
					"type": "string"
				}
			},
			"required": [
				"email",
				"password"
			],
			"type": "object",
			"additionalProperties": false
		},
		"IUserExport": {
			"properties": {
				"email": {
					"type": "string"
				},
				"role": {
					"type": "number",
					"format": "double"
				},
				"name": {
					"type": "string"
				},
				"surname": {
					"type": "string"
				},
				"patronymic": {
					"type": "string",
					"x-nullable": true
				}
			},
			"required": [
				"email",
				"role",
				"name",
				"surname"
			],
			"type": "object",
			"additionalProperties": false
		},
...
		"IRole": {
			"properties": {
				"id": {
					"type": "string"
				},
				"name_role": {
					"type": "string"
				},
				"caption_role": {
					"type": "string"
				}
			},
			"required": [
				"id",
				"name_role",
				"caption_role"
			],
			"type": "object",
			"additionalProperties": false
		},
...
	},
	"info": {
		"title": "backend",
		"version": "1.0.2",
		"license": {
			"name": "ISC"
		}
	},
	"paths": {
		"/user/register": {
			"post": {
				"operationId": "CreateUser",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"properties": {
								"token": {
									"type": "string"
								}
							},
							"required": [
								"token"
							],
							"type": "object"
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/IUser"
						}
					}
				]
			}
		},
		"/user/login": {
			"post": {
				"operationId": "LoginUser",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"properties": {
								"token": {
									"type": "string"
								}
							},
							"required": [
								"token"
							],
							"type": "object"
						}
					}
				},
				"description": "login для пользователя",
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/AuthCred"
						}
					}
				]
			}
		},
...
		"/user/userInfo": {
			"get": {
				"operationId": "GetUserByToken",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/IUserExport"
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "token",
						"required": true,
						"type": "string"
					}
				]
			}
		},
...
		"/role/{id}": {
			"get": {
				"operationId": "GetOne",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/IRole"
						}
					}
				},
				"tags": [
					"Role"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"format": "double",
						"type": "number"
					}
				]
			},
			"delete": {
				"operationId": "Delete",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/IRole"
						}
					}
				},
				"tags": [
					"Role"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/IRole"
						}
					}
				]
			}
		},
		"/role": {
			"get": {
				"operationId": "GetAll",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"items": {
								"$ref": "#/definitions/IRole"
							},
							"type": "array"
						}
					}
				},
				"tags": [
					"Role"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "Create",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/IRole"
						}
					}
				},
				"tags": [
					"Role"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/IRole"
						}
					}
				]
			},
			"patch": {
				"operationId": "Update",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/IRole"
						}
					}
				},
				"tags": [
					"Role"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/IRole"
						}
					}
				]
			}
		},
...
	},
	"produces": [
		"application/json"
	],
	"swagger": "2.0",
	"securityDefinitions": {
		"api_key": {
			"type": "apiKey",
			"name": "token",
			"description": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImFkbWluQGFkbWluIiwiaWQiOjgsInJvbGUiOiJBRE1JTiIsImlhdCI6MTYyNzI5MjAyOX0.smA9a5v47rs-oG2loD1jR5Zd5RlGqmb0XN0KWjniyKo",
			"in": "header"
		}
	},
...
}